add poi.js
poi.js contains POI class that extends Place class place.js with get_icon function that is used to get a ThemedIcon corresponding to a poi type.
A poiTypes dictionary that contains information about poi-icons for a given place type.
and a function to convert place as json to GeocodePlace.